requirements.txt
flask==3.0.0
flask-socketio==5.3.5
openai>=1.52.0
python-dotenv==1.0.0
